requirements.txt
arrow==0.17.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
monero==0.7.5
Pillow==8.1.0
psycopg2-binary==2.8.6
pysha3==1.0.2
python-dateutil==2.8.1
python-dotenv==0.15.0
qrcode==6.1
redis==3.5.3
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.22
urllib3==1.26.2
Werkzeug==1.0.1
WTForms==2.3.3
